+++ packages/Auxilium-App/src/screens/Onboarding/Slide.tsx
@@ -1,14 +1,27 @@
 import React from 'react';
-import { SafeAreaView, StyleSheet, Dimensions, View, Text } from 'react-native';
+import { SafeAreaView, StyleSheet, Dimensions, View, Text, Image } from 'react-native';
 
 const { width, height } = Dimensions.get('window');
 
 export interface SlideProps {
 	id?: number;
-	image: React.ReactNode;
+	image: any;
 	title: string;
 	description: string;
 }
+
+const Slide = ({ image, title, description }: SlideProps) => (
+	<SafeAreaView style={styles.container}>
+		<View style={styles.image}>
+			<Image
+				source={image}
+				style={{ height: 0.9 * (height / 2), width: 0.9 * width }}
+			/>
+		</View>
+		<Text style={styles.title}>{title}</Text>
+		<Text style={styles.description}>{description}</Text>
+	</SafeAreaView>
+);
 
 const styles = StyleSheet.create({
 	container: {
@@ -38,16 +51,5 @@ const styles = StyleSheet.create({
 		color: '#505050'
 	}
 });
-
-const Slide = (props: SlideProps) => {
-	const { image, title, description } = props;
-	return (
-		<SafeAreaView style={styles.container}>
-			<View style={styles.image}>{image}</View>
-			<Text style={styles.title}>{title}</Text>
-			<Text style={styles.description}>{description}</Text>
-		</SafeAreaView>
-	);
-};
 
 export default Slide;
